+++ automation/data_aggregator/__init__.py
@@ -1,0 +1,2 @@
+from .base import BaseAggregator
+from .builder import build_data_aggregator_class
