+++ automation/data_aggregator/base.py
