+++ automation/data_aggregator/builder.py
@@ -1,0 +1,4 @@
+# Write a method to build a DataAggregator class, that pulls in the relevant pieces
+
+
+def build_data_aggregator_class(manager_params)
