+++ automation/data_aggregator/local.py
