+++ automation/data_aggregator/s3.py
