+++ test/test_s3_aggregator.py
@@ -8,7 +8,7 @@ import pytest
 from localstack.services import infra
 
 from ..automation import TaskManager
-from ..automation.DataAggregator.parquet_schema import PQ_SCHEMAS
+from ..automation.data_aggregator.parquet_schema import PQ_SCHEMAS
 from .openwpmtest import OpenWPMTest
 from .utilities import (BASE_TEST_URL, LocalS3Dataset, LocalS3Session,
                         local_s3_bucket)
